Issue #OS-35 made elastic-search as separate jar
