+++ java/elastic-search/src/main/java/io/opensaber/elastic/IElasticService.java
@@ -1,7 +1,6 @@
-package io.opensaber.registry.service.impl;
+package io.opensaber.elastic;
 
 import io.opensaber.pojos.SearchQuery;
-import java.io.IOException;
 import java.util.Map;
 import org.elasticsearch.rest.RestStatus;
 
@@ -16,14 +15,14 @@ public interface IElasticService {
      * @param inputEntity - input document for adding
      * @return
      */
-    public RestStatus addEntity(String index, String id, Map<String, Object> inputEntity);
+    RestStatus addEntity(String index, String id, Map<String, Object> inputEntity);
 
     /** Reads document with respect to input osid from ES
      * @param index - ElasticSearch Index
      * @param osid - which maps to document
      * @return
      */
-    public Map<String, Object> readEntity(String index, String osid);
+    Map<String, Object> readEntity(String index, String osid);
 
     /** updates document with respect to input osid to ES
      * @param index - ElasticSearch Index
@@ -31,18 +30,18 @@ public interface IElasticService {
      * @param osid - which maps to document
      * @return
      */
-    public boolean updateEntity(String index, Map<String, Object> inputEntity, String osid);
+    boolean updateEntity(String index, Map<String, Object> inputEntity, String osid);
 
     /** deletes document with respect to input osid from ES
      * @param index - ElasticSearch Index
      * @param osid - which maps to document
      */
-    public void deleteEntity(String index, String osid);
+    void deleteEntity(String index, String osid);
 
     /** searches documents from ES based on query
      * @param index - ElasticSearch Index
      * @param searchQuery - which contains details for search
      * @return
      */
-    public Map<String, Object> search(String index, SearchQuery searchQuery);
+    Map<String, Object> search(String index, SearchQuery searchQuery);
 }
